% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{as_spark_dataframe}
\alias{as_spark_dataframe}
\title{Convert an H2O Frame to a Spark DataFrame}
\usage{
as_spark_dataframe(sc, x, name = paste(deparse(substitute(x)), collapse = ""),
  strict_version_check = TRUE)
}
\arguments{
\item{sc}{Object of type \code{spark_connection}.}

\item{x}{An \code{H2OFrame}.}

\item{name}{The name to assign the data frame in Spark.}

\item{strict_version_check}{(Optional) Setting this to FALSE does not cross check version of H2O and attempts to connect.}
}
\description{
Convert an H2O Frame to a Spark DataFrame
}

