% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_context}
\alias{h2o_context}
\title{Get or create the H2OContext.}
\usage{
h2o_context(
  sc,
  conf = NULL,
  username = NA_character_,
  password = NA_character_
)
}
\arguments{
\item{sc}{Object of type \code{spark_connection} or \code{spark_jobj}.}
}
\description{
Get or create the H2OContext.
}
