% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_h2o.R
\name{h2o_release_table}
\alias{h2o_release_table}
\title{Provide integration information for rsparkling}
\usage{
h2o_release_table()
}
\value{
Returns a data frame containing rsparkling integration information
}
\description{
Provide integration information for rsparkling
}

