% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_context}
\alias{h2o_context}
\title{Get the H2OContext. Will create the context if it has not been previously created.}
\usage{
h2o_context(x, strict_version_check = TRUE)
}
\arguments{
\item{x}{Object of type \code{spark_connection} or \code{spark_jobj}.}

\item{strict_version_check}{(Optional) Setting this to FALSE does not cross check version of H2O and attempts to connect.}
}
\description{
Get the H2OContext. Will create the context if it has not been previously created.
}

